disable line-lenth check for yamllint
