+++ phpcs.xml
@@ -9,9 +9,7 @@
     <!-- inherit rules from: -->
     <rule ref="Mimmi20CodingStandard">
         <exclude name="SlevomatCodingStandard.Functions.FunctionLength.FunctionLength"/>
-        <exclude name="SlevomatCodingStandard.Classes.ClassLength.ClassTooLong"/>
         <exclude name="SlevomatCodingStandard.Complexity.Cognitive.ComplexityTooHigh"/>
-        <exclude name="SlevomatCodingStandard.Files.FileLength.FileTooLong"/>
     </rule>
 
     <!-- Paths to check -->
